+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - RGB validation rule
 - RGBA validation rule
+- Luxembourg postal code rule
 
 ### Changed